tactic_app/dude_container_env/requirements.txt
docker==3.7
Flask
flask-cors
Flask-Login
Flask-SocketIO
Flask-SSLify
Flask-WTF
gevent-websocket==0.10.1
gunicorn
markdown
openpyxl==2.6.4
requests==2.21.0
xmltodict==0.9.2